requirements.txt
discord
discord.py
discord-py-slash-command
python_dotenv
lor_deckcodes
requests